+++ lib/convert/lexer/lexer.go
@@ -3,7 +3,7 @@ package lexer
 import (
 	"unicode/utf8"
 
-	"github.com/kijimaD/na2me/lib/token"
+	"github.com/kijimaD/na2me/lib/convert/token"
 )
 
 type Lexer struct {
