+++ lib/convert/lexer/lexer_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kijimaD/na2me/lib/token"
+	"github.com/kijimaD/na2me/lib/convert/token"
 	"github.com/stretchr/testify/assert"
 )
