+++ lib/convert/parser/parser_test.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/kijimaD/na2me/lib/ast"
-	"github.com/kijimaD/na2me/lib/lexer"
-	"github.com/kijimaD/na2me/lib/token"
+	"github.com/kijimaD/na2me/lib/convert/ast"
+	"github.com/kijimaD/na2me/lib/convert/lexer"
+	"github.com/kijimaD/na2me/lib/convert/token"
 	"github.com/stretchr/testify/assert"
 )
